app_crue_traslados/requirements.txt
django==5.2.4
openpyxl==3.1.5
psycopg2-binary==2.9.10
python-dotenv==1.1.0
